Extracted cron job to a service class
